BI_DevOps-main/requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.4

# Database
psycopg2-binary==2.9.9

# Vector storage
faiss-cpu==1.8.0
chromadb==0.5.3

# LangChain
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.14

# Embeddings
openai==1.35.3
sentence-transformers==3.0.1

# NLP utilities
beautifulsoup4==4.12.3
langdetect==1.0.9
nltk==3.8.1
tiktoken==0.7.0

# HTTP (async embedding calls)
httpx==0.27.0

# Data processing
numpy==1.26.4
pandas==2.2.2

# Monitoring
prometheus-client==0.20.0

# Testing
pytest==8.2.2
testcontainers==4.5.0
